# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Outplex Employee Hub API
  description: |
    Internal REST API for the **Outplex Employee Hub** platform.

    Covers overtime management, company store (NYT), employee marketplace,
    raffles, communications, forms, breaks scheduling, and admin operations.

    ## Authentication
    Most endpoints require a valid Supabase session cookie obtained after
    authenticating via **Slack OIDC** (`/auth/callback`).
    Dev/bootstrap endpoints use a service token in the `x-dev-token` header.

    ## Roles (least → most privileged)
    | Role | Description |
    |------|-------------|
    | `employee` | Regular employee — basic access |
    | `moderator_b1` | Supervisor — manages breaks, OT, and bath time |
    | `moderator_a1` | Senior moderator — full moderation suite |
    | `moderator` | General moderator |
    | `admin` | Full administrative access |

    ## Base URL
    All paths are relative to `https://outplexemplyeerewards.vercel.app`.

  version: 1.0.0
  contact:
    name: Outplex IT
    email: admintest@outplex.com
servers:
  - url: https://outplexemplyeerewards.vercel.app
    description: Production
...
security:
  - sessionCookie: []

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sb-pdebyqbtmnqpvepvrmiw-auth-token
      description: Supabase session cookie set after Slack OIDC authentication.
    devToken:
      type: apiKey
      in: header
      name: x-dev-token
      description: Service token for dev/bootstrap endpoints (`DEV_BOOTSTRAP_TOKEN` env var).
  responses:
    Unauthorized:
      description: Authentication required — no valid session cookie present.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
    Forbidden:
      description: Valid session exists but role is insufficient.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Forbidden
...
    Success:
      description: Operation completed successfully.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Unauthorized
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
...
    OtSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        batch_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        start_time:
          type: string
          example: '08:00'
        end_time:
          type: string
          example: '12:00'
        duration_hrs:
          type: number
        shift_label:
          type: string
          nullable: true
        status:
          type: string
          enum: [available, claimed, cancelled]
        claimed_by:
          type: string
          format: uuid
          nullable: true
        claim_kind:
          type: string
          enum: [day_off, scheduled_extension, recovery]
          nullable: true
...
paths:
...
  /api/ot/publish:
    post:
      tags: [OT]
      summary: Publish OT batch from CSV rows
      description: Creates or updates an OT batch with slots. Posts Slack notification when published.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rows, batchName, status]
              properties:
                rows:
                  type: array
                  items:
                    type: object
                  description: Parsed CSV rows from the OT upload
                batchName:
                  type: string
                batchId:
                  type: string
                  format: uuid
                  description: Provide to update an existing batch
                status:
                  type: string
                  enum: [published, draft]
      responses:
        '200':
          description: Published batch
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OtSlot'
                  batch:
                    type: object
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/ot/slots/{slotId}:
    patch:
      tags: [OT]
      summary: Update an OT slot
      description: Edits slot date, time, assignment, or status with conflict checking and user notifications.
      security:
        - sessionCookie: []
      parameters:
        - name: slotId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                start_time:
                  type: string
                  example: '08:00'
                end_time:
                  type: string
                  example: '12:00'
                shift_label:
                  type: string
                assignedUserId:
                  type: string
                  format: uuid
                claimKind:
                  type: string
                  enum: [day_off, scheduled_extension, recovery]
                status:
                  type: string
                  enum: [available, claimed, cancelled]
      responses:
        '200':
          description: Updated slot
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [OT]
      summary: Cancel an OT slot
      security:
        - sessionCookie: []
      parameters:
        - name: slotId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...